Handle new pair nav in MainScreen instead of QuickScreen
