+++ app/src/main/java/dev/arkbuilders/rate/presentation/quick/QuickScreen.kt
@@ -23,7 +23,6 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -55,7 +54,6 @@ import dev.arkbuilders.rate.domain.model.CurrencyName
 import dev.arkbuilders.rate.domain.model.PinnedQuickPair
 import dev.arkbuilders.rate.domain.model.QuickPair
 import dev.arkbuilders.rate.presentation.destinations.AddQuickScreenDestination
-import dev.arkbuilders.rate.presentation.quick.glancewidget.action.AddNewPairAction.Companion.ADD_NEW_PAIR
 import dev.arkbuilders.rate.presentation.theme.ArkColor
 import dev.arkbuilders.rate.presentation.ui.AppButton
 import dev.arkbuilders.rate.presentation.ui.AppHorDiv16
@@ -72,7 +70,6 @@ import dev.arkbuilders.rate.presentation.ui.QuickSwipeItem
 import dev.arkbuilders.rate.presentation.ui.RateSnackbarHost
 import dev.arkbuilders.rate.presentation.ui.SearchTextField
 import dev.arkbuilders.rate.presentation.utils.DateFormatUtils
-import dev.arkbuilders.rate.presentation.utils.findActivity
 import org.orbitmvi.orbit.compose.collectAsState
 import org.orbitmvi.orbit.compose.collectSideEffect
 import java.time.OffsetDateTime
@@ -89,15 +86,6 @@ fun QuickScreen(navigator: DestinationsNavigator) {
     val state by viewModel.collectAsState()
     val snackState = remember { SnackbarHostState() }
     val ctx = LocalContext.current
-    LaunchedEffect(key1 = Unit) {
-        val activity = ctx.findActivity()
-        val intent = activity?.intent
-        val createNewPair = intent?.getStringExtra(ADD_NEW_PAIR) ?: ""
-        if (createNewPair.isNotEmpty()) {
-            navigator.navigate(AddQuickScreenDestination())
-            intent?.removeExtra(ADD_NEW_PAIR)
-        }
-    }
     viewModel.collectSideEffect { effect ->
         when (effect) {
             is QuickScreenEffect.ShowSnackbarAdded ->
